refactor(tools): extract spawn-agent / time / fs-tools from index.ts (#750)
index.ts shrinks from 302 to ~75 lines, becoming a pure orchestrator.
No behavior changes — pure file split. createTimeTool and
SpawnAgentToolOptions are re-exported from index.ts so external callers
(tests) continue to import from the same path.

- spawn-agent.ts (~145 lines) — createSpawnAgentTool + SpawnAgentToolOptions
- time.ts        (~30 lines)  — createTimeTool
- fs-tools.ts    (~40 lines)  — createFsTools (renamed to avoid clash with middleware/fs.ts)
